[WIP] payment_paghiper
Adapt payment acquirer to Odoo 15.
