+++ payment_paghiper/data/paghiper.xml
@@ -1,33 +1,42 @@
 <?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data noupdate="1">
-        <record id="payment_acquirer_paghiper" model="payment.acquirer">
-            <field name="name">PagHiper</field>
-            <field name="provider">paghiper</field>
-            <field name="company_id" ref="base.main_company"/>
-            <field name="view_template_id" ref="paghiper_boleto_button"/>
-            <field name="pre_msg">
-                <![CDATA[ <p>Você vai ser redirecionado ao site da PagHiper para o boleto de pagamento.</p> ]]>
-            </field>
-            <field name="pending_msg">
-                <![CDATA[ <p>O seu pedido foi processado e está Aguardando Pagamento.</p> ]]>
-            </field>
-            <field name="done_msg">
-                <![CDATA[ <p>O seu pedido foi pago.</p> ]]>
-            </field>
-            <field name="cancel_msg">
-                <![CDATA[ <p>O seu pagamento foi cancelado.</p> ]]>
-            </field>
-            <field name="description" type="html">
-                <p>
+<odoo noupdate="1">
+  <record id="payment_acquirer_paghiper" model="payment.acquirer">
+    <field name="name">PagHiper</field>
+    <field name="provider">paghiper</field>
+    <field name="redirect_form_view_id" ref="paghiper_boleto_button" />
+    <field name="pre_msg">
+      <![CDATA[ <p>Você vai ser redirecionado ao site da PagHiper para o boleto de pagamento.</p> ]]>
+    </field>
+    <field name="pending_msg">
+      <![CDATA[ <p>O seu pedido foi processado e está Aguardando Pagamento.</p> ]]>
+    </field>
+    <field name="done_msg">
+      <![CDATA[ <p>O seu pedido foi pago.</p> ]]>
+    </field>
+    <field name="cancel_msg">
+      <![CDATA[ <p>O seu pagamento foi cancelado.</p> ]]>
+    </field>
+    <field name="description" type="html">
+      <p>
                     Muito além de pagamentos online:
                     uma infraestrutura financeira para seu negócio crescer
                 </p>
-                <ul class="list-inline">
-                    <li><i class="fa fa-check"/>eCommerce</li>
-                    <li><i class="fa fa-check"/>Botão de pagamento em cotações online</li>
-                </ul>
-            </field>
-        </record>
-    </data>
+      <ul class="list-inline">
+        <li>
+          <i class="fa fa-check" />
+          eCommerce
+        </li>
+        <li>
+          <i class="fa fa-check" />
+          Botão de pagamento em cotações online
+        </li>
+      </ul>
+    </field>
+  </record>
+
+  <record id="payment_method_paghiper" model="account.payment.method">
+    <field name="name">PagHiper</field>
+    <field name="code">paghiper</field>
+    <field name="payment_type">inbound</field>
+  </record>
 </odoo>
