+++ payment_paghiper/models/__init__.py
@@ -1,4 +1,6 @@
-from . import paghiper
+from . import payment_acquirer
+from . import payment_transaction
 from . import sale_order
 from . import account_journal
 from . import account_move
+from . import account_payment_method
