+++ payment_paghiper/views/account_journal.xml
@@ -4,7 +4,7 @@
         <field name="model">account.journal</field>
         <field name="inherit_id" ref="account.view_account_journal_form" />
         <field name="arch" type="xml">
-            <field name="outbound_payment_method_ids" position="after">
+            <field name="code" position="after">
                 <field name="receive_by_paghiper"/>
             </field>
         </field>
